requirements.txt
backtrader>=1.9.78
beautifulsoup4>=4.14
litellm>=1.0
matplotlib>=3.10
numpy>=2.4
pandas>=3.0
pydantic>=2.12
pytest>=8.0
python-dotenv>=1.0
PyYAML>=6.0
requests>=2.33
tavily-python>=0.5
transformers>=4.50
torch>=2.1
yfinance>=1.2
